# fix: replace deprecated dag.layers() with dag.multigraph_layers() for large circuit scalability